+++ packages/data/src/components/registry-provider/use-registry.js
@@ -33,7 +33,7 @@ import { Context } from './context';
  * const registry = createRegistry( {} );
  *
  * const SomeChildUsingRegistry = ( props ) => {
- *   const registry = useRegistry( registry );
+ *   const registry = useRegistry();
  *   // ...logic implementing the registry in other react hooks.
  * };
  *